Controller

+++ src/main/java/com/maltseva/servicestation/project/service/WarehouseService.java
@@ -1,6 +1,5 @@
 package com.maltseva.servicestation.project.service;
 
-
 import com.maltseva.servicestation.project.dto.WarehouseDTO;
 import com.maltseva.servicestation.project.model.ServiceStation;
 import com.maltseva.servicestation.project.model.SparePart;
@@ -12,18 +11,24 @@ import org.springframework.stereotype.Service;
 import org.webjars.NotFoundException;
 
 import java.time.LocalDateTime;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
+
+/**
+ * @author Maltseva
+ * @version 1.0
+ * @since 13.12.2022
+ */
 
 @Service
 public class WarehouseService extends GenericService<Warehouse, WarehouseDTO> {
 
     private final WarehouseRepository warehouseRepository;
     private final ServiceStationRepository serviceStationRepository;
+
     private final SparePartRepository sparePartRepository;
 
-    public WarehouseService(WarehouseRepository warehouseRepository, ServiceStationRepository serviceStationRepository,
+    public WarehouseService(WarehouseRepository warehouseRepository,
+                            ServiceStationRepository serviceStationRepository,
                             SparePartRepository sparePartRepository) {
         this.warehouseRepository = warehouseRepository;
         this.serviceStationRepository = serviceStationRepository;
@@ -39,18 +44,12 @@ public class WarehouseService extends GenericService<Warehouse, WarehouseDTO> {
     public Warehouse updateFromDTO(WarehouseDTO object, Long objectId) {
         Warehouse warehouse = warehouseRepository.findById(objectId).orElseThrow(
                 () -> new NotFoundException("Warehouse with such id = " + objectId + " not found"));
-        updateFromWarehouseDTO(object, warehouse);
-        return warehouseRepository.save(warehouse);
-    }
-
-    public void updateFromWarehouseDTO(WarehouseDTO object, Warehouse warehouse) {
         warehouse.setName(object.getName());
-        warehouse.setServiceStationID(object.getServiceStationID());
-        ServiceStation serviceStation = serviceStationRepository.findById(object.getServiceStationID()).orElseThrow(
-                () -> new NotFoundException("Service station with such id = " + object.getServiceStationID() + " not found"));
+
+        ServiceStation serviceStation = serviceStationRepository.findById(object.getServiceStationId()).orElseThrow(
+                () -> new NotFoundException("Service station with such id = " + object.getServiceStationId() + " not found"));
         warehouse.setServiceStation(serviceStation);
-        Set<SparePart> sparePartSet = new HashSet<>(sparePartRepository.findAllById(object.getSparePartSet()));
-        warehouse.setSparePartSet(sparePartSet);
+        return warehouseRepository.save(warehouse);
     }
 
     @Override
@@ -61,16 +60,26 @@ public class WarehouseService extends GenericService<Warehouse, WarehouseDTO> {
     @Override
     public Warehouse createFromDTO(WarehouseDTO newObject) {
         Warehouse newWarehouse = new Warehouse();
-        updateFromWarehouseDTO(newObject, newWarehouse);
+        newWarehouse.setName(newObject.getName());
+
+        ServiceStation serviceStation = serviceStationRepository.findById(newObject.getServiceStationId()).orElseThrow(
+                () -> new NotFoundException("Service station with such id = " + newObject.getServiceStationId() + " not found"));
+        newWarehouse.setServiceStation(serviceStation);
+
         newWarehouse.setCreatedBy(newWarehouse.getCreatedBy());
         newWarehouse.setCreatedWhen(LocalDateTime.now());
         return warehouseRepository.save(newWarehouse);
     }
 
-    public void delete(Long objectId) {
+    public void delete(Long objectId) throws ServiceException {
         Warehouse warehouse = warehouseRepository.findById(objectId).orElseThrow(
                 () -> new NotFoundException("Warehouse station with such id = " + objectId + " not found"));
-        warehouseRepository.delete(warehouse);
+        List<SparePart> sparePartList = sparePartRepository.allSparePartByWarehouseId(warehouse.getId());
+        if (sparePartList.isEmpty()) {
+            warehouseRepository.delete(warehouse);
+        } else {
+            throw new ServiceException("Невозможно удалить склад, т.к. на складе есть детали");
+        }
     }
 
     @Override
@@ -81,6 +90,10 @@ public class WarehouseService extends GenericService<Warehouse, WarehouseDTO> {
 
     @Override
     public List<Warehouse> listAll() {
-        return warehouseRepository.findAll();
+        return warehouseRepository.allWarehouse();
+    }
+
+    public List<Warehouse> getAllWarehouseByServiceStationId(Long serviceStationId) {
+        return warehouseRepository.findByServiceStationId(serviceStationId);
     }
 }
